rest-api/pyproject.toml
[project]
name = "rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "argcomplete==3.6.2",
    "black==25.1.0",
    "certifi==2025.8.3",
    "click==8.2.1",
    "datamodel-code-generator==0.33.0",
    "dnspython==2.7.0",
    "email-validator==2.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.10",
    "fastapi-cloud-cli==0.1.5",
    "genson==1.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "inflect==7.5.0",
    "isort==6.0.1",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "mypy-extensions==1.1.0",
    "numpy==2.1.1",
    "packaging==25.0",
    "pathspec==0.12.1",
    "platformdirs==4.4.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.1.0",
    "rich-toolkit==0.15.0",
    "rignore==0.6.4",
    "sentry-sdk==2.35.2",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.47.3",
    "typeguard==4.4.4",
    "typer==0.17.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
